Archive for Project_000 (The Gluttony).
Project_000 is a 30 cm tall extra-physical organism with an unnatural weight (21 kg) and a dark blue, semi-organic gel body. While lacking sensory organs or a digestive system, this entity is capable of consuming all forms of matter through a non-linear spatial absorption mechanism. Its feeding capacity is extraordinary, and it is constantly hungry. Its movements are extremely slow, but its feeding response is instantaneous.